KRACOEUS-8099:only update file name on the s2s form not on attachments page
